agenttest/requirements.txt
google-adk
opentelemetry-instrumentation-google-genai
python-dotenv
